Rename upload to publish
